Versión 1.0 lista.

+++ Dolphin Space Hunter/Assets/_Scripts/Shooting_Scripts/DifficultySettingsContainer.cs
@@ -13,7 +13,7 @@ public class DifficultySettingsContainer : MonoBehaviour
     #region Variables
     public int maxAmmunition;
     public int wavesPerBoss; 
-    public int shieldRechargeRate; //Factor que indica cuantos frames agitar el telefono para que aumente el escudo en 1 unidad
+    public float shieldRechargeRate; //Factor que indica cuantos frames agitar el telefono para que aumente el escudo en 1 unidad
     public int damagePerAsteroid;
     public float asteroidSpawnRate;
     public float minAsteroidSpeed, maxAsteroidSpeed;
